pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.0.1"
flask = "^2.2.2"
peewee = "^3.15.3"
parsedatetime = "^2.6"
loguru = "^0.6.0"
pytz = "^2022.4"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
